+++ R/print.curtailment_simon.R
@@ -1,0 +1,1 @@
+print.curtailment_simon <- function(x, ...) print(x$all.des)